move IdentityHasher into its own file
