+++ src/serde/mod.rs
@@ -3,6 +3,7 @@ mod de;
 mod de_br;
 mod de_tree;
 mod errors;
+mod identity_hash;
 mod incremental;
 mod object_cache;
 mod parse_atom;
@@ -19,6 +20,7 @@ mod test;
 pub use de::node_from_bytes;
 pub use de_br::{node_from_bytes_backrefs, node_from_bytes_backrefs_record};
 pub use de_tree::{parse_triples, ParsedTriple};
+pub use identity_hash::RandomState;
 pub use incremental::{Serializer, UndoState};
 pub use object_cache::{serialized_length, treehash, ObjectCache};
 pub use ser::{node_to_bytes, node_to_bytes_limit};
